:white_check_mark: test(DepartmentServiceTest): 임의의 급여 인상 조건 테스트 이후 업데이트 로직 서비스 테스트 완료

+++ src/main/java/com/api/ecoassignment/domain/department/dto/response/DepartmentAndEmployeeDto.java
@@ -13,6 +13,6 @@ public class DepartmentAndEmployeeDto {
     private Long employeeId;
     private String departmentName;
     private BigDecimal salary;
-    private Integer minSalary;
-    private Integer maxSalary;
+    private BigDecimal minSalary;
+    private BigDecimal maxSalary;
 }
